Refactor namespace constants
