+++ pkg/constants/constants.go
@@ -52,6 +52,9 @@ const VerrazzanoInstallNamespace = "verrazzano-install"
 // VerrazzanoSystemNamespace is the system namespace for Verrazzano
 const VerrazzanoSystemNamespace = "verrazzano-system"
 
+// VerrazzanoCAPINamespace is the system namespace for Cluster API resources
+const VerrazzanoCAPINamespace = "verrazzano-capi"
+
 // VerrazzanoMultiClusterNamespace is the multi-cluster namespace for Verrazzano
 const VerrazzanoMultiClusterNamespace = "verrazzano-mc"
