+++ tests/e2e/verify-install/capi/capi_test.go
@@ -7,9 +7,9 @@ import (
 	"fmt"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/verrazzano/verrazzano/pkg/constants"
 	"github.com/verrazzano/verrazzano/pkg/k8sutil"
 	"github.com/verrazzano/verrazzano/pkg/vzcr"
-	"github.com/verrazzano/verrazzano/platform-operator/constants"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg/test/framework"
 	"time"
@@ -59,7 +59,7 @@ func WhenCapiInstalledIt(description string, f func()) {
 		if err != nil {
 			AbortSuite(fmt.Sprintf("Failed to check Verrazzano version 1.6.0: %v", err))
 		}
-		isComponentStatusEnabled := vzcr.IsComponentStatusEnabled(inClusterVZ, constants.VerrazzanoCAPINamespace
+		isComponentStatusEnabled := vzcr.IsComponentStatusEnabled(inClusterVZ, constants.VerrazzanoCAPINamespace)
 		if isMinimumK8sVersion && isCAPISupported && (isCAPIEnabled && isComponentStatusEnabled) {
 			f()
 		} else {
